Update README.md
Changed this "scope": "openid profile"	for this "scope": "openid name email picture"	

Modified this Or just parse the hash (if loginOption.scope is not openid profile, then the profile will only contains the user_id):  for this Or just parse the hash (if loginOption.scope is `openid`, then the profile will only contain the `user_id`):
Taking this into consideration: " 'openid': It will return, not only the access_token, but also an id_token which is a Json Web Token (JWT). The JWT will only contain the user id (sub claim). You can use constant ParemterBuilder.SCOPE_OPENID."